requirements.txt
streamlit
torch
torchvision
torchaudio
numpy
opencv-python
Pillow